+++ imports/api/attachments/server/methods.js
@@ -1,5 +1,11 @@
 import { Attachments } from "/imports/api/attachments/attachments.js";
 import { checkLoggedIn } from "/imports/api/permissions/permissions";
+import fs from "fs";
+import { checkCanWriteTask } from "/imports/api/permissions/permissions";
+
+const bound = Meteor.bindEnvironment(callback => {
+  return callback();
+});
 
 Attachments.methods.remove = new ValidatedMethod({
   name: "attachments.remove",
@@ -51,7 +57,6 @@ Attachments.methods.remove = new ValidatedMethod({
   }
 });
 
-
 Attachments.methods.restore = new ValidatedMethod({
   name: "attachments.restore",
   validate: new SimpleSchema({
@@ -63,5 +68,46 @@ Attachments.methods.restore = new ValidatedMethod({
     checkLoggedIn();
 
     // TODO : Implement method
+  }
+});
+
+Attachments.methods.clone = new ValidatedMethod({
+  name: "attachments.clone",
+  validate: new SimpleSchema({
+    attachmentId: { type: String, optional: true },
+    projectId: { type: String, optional: true },
+    taskId: { type: String, optional: true }
+  }).validator(),
+  run({ attachmentId, projectId, taskId }) {
+    checkLoggedIn();
+    checkCanWriteTask(taskId);
+
+    const attachment = Attachments.findOne({ _id: attachmentId });
+    const userId = Meteor.userId()
+    fs.readFile(attachment.path, function(error, data) {
+      bound(() => {
+        if (error) {
+          throw error;
+        } else {
+          Attachments.write(
+            data,
+            {
+              fileName: attachment.name,
+              type: attachment.type,
+              meta: {
+                projectId: projectId,
+                taskId: taskId,
+                createdBy: userId
+              }
+            },
+            function(writeError, fileRef) {
+              if (writeError) {
+                throw writeError;
+              }
+            }
+          );
+        }
+      });
+    });
   }
 });